requirements.txt
Flask
APScheduler
numpy
pandas
twilio
Adafruit-DHT
RPi.GPIO
gpiozero
tensorflow
picamera
Pillow